Create printers for roster formatting, code refactoring
