+++ api/formatter/formats/__init__.py
@@ -1,0 +1,4 @@
+from .format_printer import FormatPrinter
+from .rus import RussianTournamentsPrinter
+from .default_printer import DefaultPrinter
+from .wtc_printer import WTCPrinter
