+++ api/formatter/formats/format_printer.py
@@ -1,0 +1,7 @@
+from ..rosterview import RosterView
+
+
+class FormatPrinter:
+    @staticmethod
+    def print(roster: RosterView) -> str:
+        raise NotImplementedError
